Chnage data type

+++ app/src/main/java/com/mad/p03/np2020/routine/Home.java
@@ -225,7 +225,7 @@ public class Home extends AppCompatActivity implements OnSectionListener, MyData
     }
 
     @Override
-    public void onSectionDelete(int position) {
+    public void onSectionDelete(long position) {
 
     }
